+++ src/elfpy/policies/init_lp.py
@@ -4,7 +4,7 @@ Special reserved user strategy that is used to initialize a market with a desire
 # pylint: disable=duplicate-code
 # pylint: disable=too-many-arguments
 
-from elfpy.strategies.basic import BasicPolicy
+from elfpy.policies.basic import BasicPolicy
 from elfpy.pricing_models import ElementPricingModel
