+++ src/elfpy/policies/single_long.py
@@ -3,7 +3,7 @@ User strategy that opens a long position and then closes it after a certain amou
 """
 # pylint: disable=too-many-arguments
 
-from elfpy.strategies.basic import BasicPolicy
+from elfpy.policies.basic import BasicPolicy
 
 
 class Policy(BasicPolicy):
