+++ src/elfpy/policies/single_lp.py
@@ -6,7 +6,7 @@ User strategy that adds base liquidity and doesn't remove until liquidation
 # pylint: disable=duplicate-code
 # pylint: disable=too-many-arguments
 
-from elfpy.strategies.basic import BasicPolicy
+from elfpy.policies.basic import BasicPolicy
 
 
 class Policy(BasicPolicy):
