+++ tests/utils/test_parse_config_success_data.toml
@@ -30,7 +30,7 @@ num_trading_days = 180  # should be <= pool_duration
 num_blocks_per_day = 7200 # 24 * 60 * 60 / 12 = 12 second block time
 token_duration = 180  # in days
 precision = 64 # 64 is max precision (and the default for numpy)
-agent_policies = ["single_long"] # specify a list of trading strategies by name
+agent_policies = ["single_long"] # specify a list of trading policies by name
 random_seed = 123 # to be passed to a rng
 verbose = false # toggle verbose output level
 shuffle_users = true
